ldap

+++ config/ldap.php
@@ -32,7 +32,7 @@ return [
             'hosts' => [env('LDAP_HOST', '127.0.0.1')],
             'username' => env('LDAP_USERNAME'),
             'password' => env('LDAP_PASSWORD'),
-            'port' => env('LDAP_PORT', 389),
+            'port' => (int) env('LDAP_PORT', 389),
             'base_dn' => env('LDAP_BASE_DN'),
             'timeout' => env('LDAP_TIMEOUT', 5),
             'use_ssl' => env('LDAP_SSL', false),
